Merge http://lia64.bkbits.net/linux-ia64-release-2.6.9 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/09/30 16:31:06+00:00 eranian
[IA64] minor fix to perfmon

change the return value of PFM_CREATE_CONTEXT from
EAGAIN to ENOMEM when the sampling buffer size is
larger than the RLIMIT_MEMLOCK limit of the task.

signed-off-by: Stephane Eranian <eranian@hpl.hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/09/30 16:27:05+00:00 eranian
[IA64] perfmon2 fix for TASK_TRACED

fix a problem in pfm_check_task_state() and pfm_task_incompatible()
which was caused by the introduction of the new TASK_TRACED state.
Tool would fail to attach to a process,i.e., PFM_LOAD_CONTEXT would
fail. With the fix perfmon now accepts to operate on tasks which are
in either TASK_STOPPED or TASK_TRACED state. The problem was tracked
down by Alex Williamson from HP who also submitted the patch.

signed-off-by: Stephane Eranian <eranian@hpl.hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/09/29 18:27:52+00:00 tony.luck
[IA64] SMP systems may not have SRAT, still need to mark node0 online.

This only causes a problem in the -mm tree now, but Zou Nan hai and
Shaohua Li sent me this fix so that SMP systems compiled with the
generic kernel (which turns on NUMA and CPUSETS in -mm) won't hit
a BUG_ON in kernel/cpuset.c guarantee_online_mems() which is called
from acpi_early_init()

Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/09/29 17:03:07+00:00 jbarnes
[IA64] defconfig for Intel bigsur

Here's a defconfig that comes up on my box and has a reasonable set of modules
enabled.

Signed-off-by: Jesse Barnes <jbarnes@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

2004/09/29 16:58:05+00:00 jbarnes
[IA64-SGI] sn2: serialize access to PROM chips

If we read and write the PROM chips at the same time, as might happen at boot
when salinfo extracts MCA records and a user is checking the PROM revision
in /proc/sgi_prominfo, an MCA might occur, since the PROM chips can't be
accessed that way.  This patch fixes the problem for systems with new PROMs
(>= 3.50) by using the SAL to do PROM reads.

Signed-off-by: Jesse Barnes <jbarnes@sgi.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>

BKrev: 415c4540al9SqFmQQQR5lOliPrqqkA