[NET]: Fix fddi_statistics for 64-bit
 There is a problem with "struct fddi_statistics" for 64-bit systems.
The starting members of the struct are expected to correspond to the
respective members of "struct net_device_stats" (drivers for FDDI
devices return "struct fddi_statistics" in the response to the
get_stats() call of "struct net_device").  Unfortunately, due to using
different types (u32 vs ulong) they do not.  "struct net_device_stats"
is a public interface and as a result, bogus results are retrieved,
e.g. for /proc/net/dev.

 Here is my proposal to address the problem.  I think there is no
point in duplicating the layout of "struct net_device_stats" in
"struct fddi_statistics" as the former can simply be included as a
member avoiding this problem and actually any possible discrepancy in
the future.  This also preserves the layout of the structure for
32-bit systems.

Signed-off-by: Maciej W. Rozycki <macro@linux-mips.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 4162f6291mYUARaajH3ZDYq53MfGfg