[PATCH] x86_64: fix profile_pc
This fixes profile_pc to work properly on x86-64 and not crash.

It does now a simple backtrace to the caller of the spin lock without
requiring a frame pointer for this.

Frame pointer support has been dropped because it never worked.

There is still a small race window, but the only way to avoid it would be
to rewrite kernel/spinlock.c in assembler again.  The race will account a
profile tick the the parent of the spinlock caller.

Signed-off-by: Andi Kleen <ak@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41640ba6CdKyFL39n0cjSNpBg8I80A