[PATCH] revert writeback threshold changes
The post-2.6.8 recent tweaks to the page writeback thresholding code seems to
be doing the wrong thing for small unmapped_ratios - nathans had a few
problems with it.  The problem it was trying to fix is minor and fairly
theoretical anyway, so back it out to the old code.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 416dffe57KkxRYEbyOjvdtPtoLUhbg