[PATCH] ppc64:  fix smp_startup_cpu for cpu hotplug
This change is needed in order to allow cpus to be onlined after
boot.  This used to work but the declaration of
pseries_secondary_smp_init in this file was changed in Ben's big
cleanup patch a while back, so the cpu would start at a bad address.

Signed-off-by: Nathan Lynch <nathanl@austin.ibm.com>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41729bf4dLr_iUHVxxyesAp2tiIWpw