[PATCH] invalidate page race fix
invalidate_inode_pages() and invalidate_inode_pages2() can mark pages not
uptodate while read() is trying to read from them.  This is interpreted as
an I/O error.

Fix that by teaching the invalidate code to leave the page alone if someone
else has a ref on it.

Signed-off-by: Jens Axboe <axboe@suse.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4174aca2ocZwQ_22QLBHXsj0hDWUWw