[PATCH] sh: fix EMBEDDED_RAMDISK with O=
The following fixes EMBEDDED_RAMDISK to work with O=.  The problem was that
we couldn't find the linker script, since we needed to specify the patch to
the source tree for it.  I've tested this with the ramdisk set to both
'ramdisk.gz' and '../ramdisk.gz'.

Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4174adcd0E5elarBwLjF0mnYIkXLgQ