[PATCH] SELinux: allow all filesystems to specify fscreate mount  option
The patch below allows all types of filesystems to specify the fscreate
mount option (which is used to specify the security context of the
filesystem itself).  This was previously only available for filesystems
with full xattr security labeling, but is also potentially required for
filesystems with e.g.  psuedo xattr labeling such as devpts and tmpfs.

An example of use is to specify at mount time the fs security context of a
tmpfs filesystem, overriding the default specified in policy for that
filesystem.

This patch has been in the Fedora kernel for some weeks with no problems.

Signed-off-by: James Morris <jmorris@redhat.com>
Signed-off-by: Stephen Smalley <sds@epoch.ncsc.mil>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41752ffe7wJOiTHitcJO4PVvSXxpIA