[PATCH] I2C: lm87 driver ported to Linux 2.6
This is my port of the lm87 driver to Linux 2.6. It is based on the
preliminary work of Jeff Oliver. I then significantly improved the code,
added functionality, tested the whole thing on a real motherboard, fixed
a couple remaining bugs, and here we are.

I'll port a number of improvements and fixes back to the 2.4 version of
the driver after lm_sensors 2.8.8 is released (i.e. soon).

Signed-off-by: Jean Delvare <khali@linux-fr.org>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>

BKrev: 4175942dqSAg96coADp_8YNMLfEoRA