[PATCH] v4l: avoid using struct file ptrs in video-buf
This patch makes the video-buf helper module pass through a void pointer
instead of a struct file pointer, that makes the code also usable when no
file pointer is available.  This is needed for when using the video-buf
infrastructure to manage DMA buffers for DVB cards.

The file pointer was used by the videobuf_queue_ops callbacks to get the
drivers private data via file->private_data, now a pointer to the drivers's
private data can be passed directly.

Adaptions of the drivers follow with separate patches.

Signed-off-by: Gerd Knorr <kraxel@bytesex.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41768196E6i6bbiUnhTg9OFubLPWHA