[PATCH] DVB/V4L dependency fix
The problem is that dvb is way to much work-in-progress right now, it's
almost impossible to build something which works on both mainline kernel +
latest dvb cvs.  Hope at least the interfaces are settled soon, for now I
gave up on getting it work in mainline in short-term.

The patch below compiles the DVB card configuration code only when
VIDEO_CX88_DVB is enabled (which is commented right now in Kconfig due to
being broken).

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 417681cfeddXm6Bn4XsA4SKf1Q-rFg