[PATCH] gen_init_cpio uses external file list
This patch makes gen_init_cpio generate the initramfs_data.cpio from a file
which contains a list of entries: file, dir, nod.  I swapped the order of
filename/location for the file arguments so that it would be more uniform
with the dir and node tyes.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 417682d8HQOCqOtOE_jdNsKtiXGfHQ